gatekeeper utils for bookmark user and goals
