+++ client/components/index.js
@@ -11,3 +11,4 @@ export {default as Landing} from './landing'
 export {default as AllGoals} from './all-goals'
 export {default as SingleGoal} from './single-goal'
 export {default as Blocked} from './blocked'
+export {default as ErrorMessage} from './error-message'
