+++ server/api/bookmarks.js
@@ -1,7 +1,16 @@
 const router = require('express').Router()
 const {Bookmark} = require('../db/models')
 // const { ChromeMarks } = require('../../bg')
+const {checkIfAdmin} = require('../../utils')
 module.exports = router
+
+router.use((req, res, next) => {
+  if (req.user && checkIfAdmin(req.user)) {
+    next()
+  } else {
+    res.status(401).send('ACCESS DENIED')
+  }
+})
 
 // get all, get by id, update by id, delete by id, and create routes
 router.get('/:id', async (req, res, next) => {
